// frontend/src/api-client/models/PatientDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PatientDto = {
    id?: string;
    tenantId?: string;
    externalId?: string;
    cccd?: string;
    fullNameVi?: string;
    dateOfBirth?: string;
    gender?: string;
    phone?: string;
    email?: string;
    addressLine?: string;
    city?: string;
    district?: string;
    ward?: string;
    nationality?: string;
    ethnicity?: string;
    isActive?: boolean;
    avatarUrl?: string;
    createdAt?: string;
    updatedAt?: string;
};